Refactor ChatArea into smaller components
